Shuffle where traversable lives
