+++ src/Data/Vector/Linear.hs
@@ -34,7 +34,6 @@ module Data.Vector.Linear
   ) where
 
 import qualified Data.Functor.Linear as Data
-import qualified Data.Traversable.Linear as Data
 import Data.Proxy
 import Data.Type.Equality
 import Data.Vector (Vector)
